Update readme + cleanup
