+++ README.md
@@ -21,6 +21,8 @@ We wrote an article on medium that detailed our approach. You can read it [here]
 
 ### Training Data
 
+**Outdated** (you have to use convert_old_format.py to use current code, now all the informations are in a pickle file)
+
 The training data (7600+ labeled images) can be downloaded [here](https://www.dropbox.com/s/24x9b6kob5c5847/training_data.zip?dl=0)
 
 There are two folders:
@@ -44,7 +46,7 @@ For installation, see section **Installation**.
 
 ## Autonomous mode
 
-0. Compile and upload the code on the arduino
+0. Compile and upload the code on the Arduino
 ```
 cd arduino/
 make
@@ -78,30 +80,32 @@ python command/python/teleop_client.py
 python command/python/teleop_server.py -v my_video
 ```
 2. Convert the recorded video from h264 to mp4 using ffmpeg or [MP4Box](https://gpac.wp.imt.fr/mp4box/)
-3. Split the video into a sequence of images (please change the paths in the script)
 ```
-python -m opencv.split_video -i video.mp4
+MP4Box -add video.h264 video.mp4
 ```
-You have to press enter to pass to the next frame
 
-4. Label the data using the labeling tool (again please change the paths in the script)
+3. Split the video into a sequence of images
 ```
-python -m train.label_images
+python -m train.split_video -i video.mp4 --no-display -o path/output/folder
+```
+
+4. Label the data using the labeling tool
+```
+python -m train.label_images -i path/to/input/folder -o path/to/output/folder
 ```
 To label an image, you have to click on the center of line in the displayed image.
 If the image do not contain any line, or if you want to pass to the next frame, press any key.
 
-5. (optional) Augment the dataset using the augmentDataset() function in train/train.py
-6. Train the neural network (again please change the paths in the script)
+5. Train the neural network (again please change the paths in the script)
 ```
-cd train/ (enter train folder)
-python train.py
+python -m train.train -f path/input/folder
 ```
-The best model (lowest error on the validation data) will be saved as *mlp_model.npz*.
+The best model (lowest error on the validation data) will be saved as *mlp_model_tmp.npz*.
 
 ### Installation
 
 #### Recommended : Use an image with everything installed
+
 0. You need a micro sd card (warning, all data on that card will be overwritten)
 
 1. Download the image [here](https://drive.google.com/open?id=0Bz4VOC2vLbgPTl9LZzNNcnBCWUU)
