+++ train/__init__.py
@@ -1,4 +1,3 @@
 from __future__ import absolute_import
 
-from .train import preprocessImage, WIDTH, HEIGHT, loadPytorchNetwork
-from .utils import loadVanillaNet
+from .utils import preprocessImage, loadVanillaNet, loadPytorchNetwork, loadDataset
