+++ train/train.py
@@ -5,111 +5,28 @@ Train a neural network to detect a black&white line
 from __future__ import print_function, division, absolute_import
 
 import argparse
-import os
 import time
-import pickle as pkl
 
-import cv2
 import numpy as np
 import torch as th
 import torch.utils.data
 import torch.nn as nn
-import torch.nn.functional as F
 from torch.autograd import Variable
-from sklearn.model_selection import train_test_split
 
-from constants import WIDTH, HEIGHT, INPUT_DIM, SPLIT_SEED
-from .utils import preprocessImage, CosineAnnealingLR
+from constants import SPLIT_SEED
+from .utils import preprocessImage, saveToNpz, loadDataset
 from .models import MlpNetwork, ConvolutionalNetwork
 
 evaluate_print = 1  # Print info every 1 epoch
 VAL_BATCH_SIZE = 256
 
+
 # TODO: change std of weights initialization
 
 
-def loadPytorchNetwork(model_name="mlp_model_tmp"):
-    if '.pth' in model_name:
-        model_name = model_name.split('.pth')[0]
-    model = MlpNetwork(INPUT_DIM)
-    model.load_state_dict(th.load(model_name + '.pth'))
-    model.eval()
-    return model
-
-
-def saveToNpz(model, output_name="mlp_model_tmp"):
-    """
-    :param model: (PyTorch Model)
-    """
-    np.savez(output_name, *[p.data.numpy().T for _, p in model.named_parameters()])
-
-def loadDataset(seed=42, folder='cropped', split=True, augmented=True):
-    """
-    Load the training images and preprocess them
-    :param seed: (int) seed for pseudo-random generator
-    :param folder: (str) input folder
-    :param split: (bool) Whether to split the dataset into 3 subsets (train, validation, test)
-    :param augmented: (bool)
-    :return:
-    """
-
-    with open('{}/infos.pkl'.format(folder), 'rb') as f:
-        images_dict = pkl.load(f)['images']
-
-    images = list(images_dict.keys())
-    images.sort()
-    images_path = []
-
-    # TODO: check channel order (BGR or RGB)
-    tmp_im = cv2.imread('{}/{}.jpg'.format(folder, images_dict[images[0]]['output_name']))
-    height, width, _ = tmp_im.shape
-    n_images = len(images)
-    if augmented:
-        images_path_augmented = []
-        n_images *= 2
-
-    X = np.zeros((n_images, INPUT_DIM), dtype=np.float32)
-    y = np.zeros((n_images,), dtype=np.float32)
-
-    print("original_shape=({},{})".format(width, height))
-    print("resized_shape=({},{})".format(WIDTH, HEIGHT))
-    # factor = width / WIDTH
-
-    for idx, name in enumerate(images):
-        x_center, y_center = images_dict[name]['label']
-        # TODO: check the formula below (if this changes, it must be changed in image_processing.py too)
-        y[idx] = x_center / width
-
-        path = images_dict[name]['output_name']
-        image_path = '{}/{}.jpg'.format(folder, path)
-        im = cv2.imread(image_path)
-        X[idx, :] = preprocessImage(im, WIDTH, HEIGHT)
-        images_path.append(path + '.jpg')
-        if augmented:
-            horizontal_flip = cv2.flip(im, 1)
-            X[len(images) + idx, :] = preprocessImage(horizontal_flip, WIDTH, HEIGHT)
-            images_path_augmented.append(path + '.jpg')
-            y[len(images) + idx] = (width - x_center) / width
-
-    if augmented:
-        images_path += images_path_augmented
-
-    print("Input tensor shape: ", X.shape)
-
-    if not split:
-        return X, y, images_path
-
-    # for CNN
-    # X = X.reshape((-1, WIDTH, HEIGHT, 3))
-    # X = np.transpose(X, (0, 3, 2, 1))
-    print(X.shape)
-    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.4, random_state=seed)
-    X_val, X_test, y_val, y_test = train_test_split(X_test, y_test, test_size=0.5, random_state=seed)
-
-    return X_train, y_train, X_val, y_val, X_test, y_test
-
-
-def main(folder, num_epochs=1000, batchsize=1, learning_rate=0.0001, seed=42, cuda=False):
+def main(folder, num_epochs=1000, batchsize=1,
+         learning_rate=0.0001, seed=42, cuda=False,
+         load_model=""):
     """
     :param folder: (str)
     :param num_epochs: (int)
@@ -117,17 +34,17 @@ def main(folder, num_epochs=1000, batchsize=1, learning_rate=0.0001, seed=42, cu
     :param learning_rate: (float)
     :param seed: (int)
     :param cuda: (bool)
+    :param load_model: (str) path to a saved model
     """
     # Load the dataset
     print("Loading data...")
 
-    X_train, y_train, X_val, y_val, X_test, y_test = loadDataset(folder=folder, seed=SPLIT_SEED)
+    X_train, y_train, X_val, y_val, X_test, y_test = loadDataset(folder=folder, split_seed=SPLIT_SEED)
     # Seed the random generator
     np.random.seed(seed)
     th.manual_seed(seed)
     if cuda:
         th.cuda.manual_seed(seed)
-
 
     kwargs = {'num_workers': 1, 'pin_memory': False} if cuda else {}
 
@@ -145,13 +62,15 @@ def main(folder, num_epochs=1000, batchsize=1, learning_rate=0.0001, seed=42, cu
                                           batch_size=VAL_BATCH_SIZE, shuffle=False, **kwargs)
 
     test_loader = th.utils.data.DataLoader(th.utils.data.TensorDataset(X_test, y_test),
-                                          batch_size=VAL_BATCH_SIZE, shuffle=False, **kwargs)
+                                           batch_size=VAL_BATCH_SIZE, shuffle=False, **kwargs)
 
     input_dim = X_train.shape[1]
     model = MlpNetwork(input_dim, n_hidden=[20, 4], drop_p=0.6)
     model_name = "mlp_model_tmp"
     # model_name = "cnn__model_tmp"
     # model = ConvolutionalNetwork()
+    if load_model != "":
+        model.load_state_dict(th.load(load_model))
 
     if cuda:
         model.cuda()
@@ -159,12 +78,13 @@ def main(folder, num_epochs=1000, batchsize=1, learning_rate=0.0001, seed=42, cu
     # L2 penalty
     # weight_decay = 1e-4
     weight_decay = 0
+    # Optimizers
     # optimizer = th.optim.Adam(model.parameters(), lr=learning_rate, weight_decay=weight_decay)
     optimizer = th.optim.SGD(model.parameters(), lr=learning_rate,
                              momentum=0.9, weight_decay=weight_decay, nesterov=True)
-    # scheduler = th.optim.lr_scheduler.StepLR(optimizer, step_size=300, gamma=0.1)
+    # Learning rate schedulers
+    # scheduler = th.optim.lr_scheduler.StepLR(optimizer, step_size=40, gamma=0.5)
     # scheduler = th.optim.lr_scheduler.MultiStepLR(optimizer, milestones=[30,80], gamma=0.1)
-    # scheduler =  CosineAnnealingLR(optimizer, T_max=10, eta_min=0.05)
 
     loss_fn = nn.MSELoss(size_average=False)
     # loss_fn = nn.SmoothL1Loss(size_average=False)
@@ -182,8 +102,10 @@ def main(folder, num_epochs=1000, batchsize=1, learning_rate=0.0001, seed=42, cu
         # Full pass on training data
         # Update the model after each minibatch
         for inputs, targets in train_loader:
+            # Move variables to gpu
             if cuda:
                 inputs, targets = inputs.cuda(), targets.cuda()
+            # Convert to pytorch variabless
             inputs, targets = Variable(inputs), Variable(targets)
             optimizer.zero_grad()
             predictions = model(inputs)
@@ -198,6 +120,7 @@ def main(folder, num_epochs=1000, batchsize=1, learning_rate=0.0001, seed=42, cu
         for inputs, targets in val_loader:
             if cuda:
                 inputs, targets = inputs.cuda(), targets.cuda()
+            # Set volatile to True because we don't need to compute gradient
             inputs, targets = Variable(inputs, volatile=True), Variable(targets)
             predictions = model(inputs)
             loss = loss_fn(predictions, targets)
@@ -207,9 +130,10 @@ def main(folder, num_epochs=1000, batchsize=1, learning_rate=0.0001, seed=42, cu
         # Save the new best model
         if val_error < best_error:
             best_error = val_error
+            # Move back weights to cpu
             if cuda:
                 model.cpu()
-
+            # Save as pytorch (pth) and numpy file (npz)
             th.save(model.state_dict(), best_model_path)
             saveToNpz(model, model_name)
 
@@ -240,15 +164,15 @@ def main(folder, num_epochs=1000, batchsize=1, learning_rate=0.0001, seed=42, cu
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Train a line detector')
+    parser.add_argument('-f', '--folder', help='Training folder', default="augmented_dataset", type=str, required=True)
     parser.add_argument('--num_epochs', help='Number of epoch', default=50, type=int)
     parser.add_argument('-bs', '--batchsize', help='Batch size', default=4, type=int)
     parser.add_argument('--seed', help='Random Seed', default=42, type=int)
     parser.add_argument('--no-cuda', action='store_true', default=False, help='disables CUDA training')
-
-    parser.add_argument('-f', '--folder', help='Training folder', default="augmented_dataset", type=str)
+    parser.add_argument('--load_model', help='Start from a saved model', default="", type=str)
     parser.add_argument('-lr', '--learning_rate', help='Learning rate', default=1e-4, type=float)
     args = parser.parse_args()
 
     args.cuda = not args.no_cuda and th.cuda.is_available()
     main(folder=args.folder, num_epochs=args.num_epochs, batchsize=args.batchsize,
-         learning_rate=args.learning_rate, cuda=args.cuda, seed=args.seed)
+         learning_rate=args.learning_rate, cuda=args.cuda, seed=args.seed, load_model=args.load_model)
